requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Serial communication
pyserial==3.5
pyserial-asyncio==0.6

# Utilities
pyyaml==6.0.1
structlog==24.1.0

# Development
ruff==0.2.1
mypy==1.8.0
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
